build: Fix undefined behavior in left shift
Spotted by UBSan.

+++ p11-kit/rpc-message.c
@@ -636,7 +636,7 @@ p11_rpc_buffer_encode_uint32 (unsigned char* data,
 uint32_t
 p11_rpc_buffer_decode_uint32 (unsigned char* ptr)
 {
-	uint32_t val = ptr[0] << 24 | ptr[1] << 16 | ptr[2] << 8 | ptr[3];
+	uint32_t val = (uint32_t) ptr[0] << 24 | ptr[1] << 16 | ptr[2] << 8 | ptr[3];
 	return val;
 }
